Refactor to use packedscene instead of node type
This refactors the projectile effect system to use packed scenes
instead of the base class node type ProjectileEffect.

This is due to the requirement for these scenes to be instantiated for
each projectile, not a shared resource. Without this change, effects
that modify the scene tree (such as HUGE LASER) do not work.